requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.0
celery>=5.3.0
playwright>=1.40.0
scikit-learn>=1.3.0
spacy>=3.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
email-validator>=2.1.0
geopy>=2.4.0